Refactor tax rate handling and store rates as a post type
Our tax rates are different from the rates expected by WooCommerce
because Washington state will vary on address and postal code. We
don't have a connection for 4 digit zip lookup, but we do have an
API that provides tax information based on a full address.

When a rate is obtained based on an address, city, and postal code,
it is stored in cache and as a custom post type. As long as the
customer's address stays the same, then the tax will remain
untouched.

Future work will need to be done to ensure that address changes
work as expected.